// cg-web/src/types/device.ts
export type DeviceModelType = {
  type?: DeviceModelTypeEnum;
  specs?: any;
  name?: string;
};

export type DevicePropsType = {
  id?: string;
  name?: string;
  groupName?: string;
  dataType?: DeviceModelType;
  properties?: DevicePropsType[];
  component?: string;
  showHistory?: boolean;
  sourceId?: string;
};